# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MiniMarket Plus API",
    "description": "Backend para gestion de inventario, ventas, carrito y usuarios con seguridad JWT.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/productos/{id}": {
      "get": {
        "tags": [
          "Producto"
        ],
        "summary": "Obtiene un producto por id",
        "description": "Retorna un producto individual como DTO, incorporando enlaces HATEOAS hacia el recurso actual, la coleccion, la categoria y el carrito.",
        "operationId": "obtenerProductoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Error interno del servidor"
          },
          "200": {
            "description": "Producto encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProductoDTO"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Producto"
        ],
        "summary": "Actualiza precio, stock o categoria de un producto",
        "description": "Actualiza los datos de un producto existente. Requiere rol administrador o jefe de turno.",
        "operationId": "actualizarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "200": {
            "description": "Producto actualizado correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProductoDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Producto"
        ],
        "summary": "Elimina un producto",
        "description": "Elimina un producto existente del catalogo. Requiere rol administrador.",
        "operationId": "eliminarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "204": {
            "description": "Producto eliminado correctamente"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/productos": {
      "get": {
        "tags": [
          "Producto"
        ],
        "summary": "Lista productos con enlaces HATEOAS",
        "description": "Retorna una coleccion paginada de productos representados como DTO y enriquecidos con enlaces self, collection, disponibles y stock-bajo.",
        "operationId": "listarProductos",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Numero de pagina a consultar, comenzando desde 0",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Cantidad de registros por pagina",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Productos consultados correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelProductoDTO"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "tags": [
          "Producto"
        ],
        "summary": "Crea un producto",
        "description": "Registra un nuevo producto en el catalogo. Requiere rol administrador o jefe de turno.",
        "operationId": "guardarProducto",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "200": {
            "description": "Producto creado correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProductoDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/productos/stock-bajo": {
      "get": {
        "tags": [
          "Producto"
        ],
        "summary": "Lista productos que requieren reposicion",
        "description": "Retorna productos cuyo stock es menor o igual al minimo informado. Requiere rol administrador o jefe de turno.",
        "operationId": "listarProductosConStockBajo",
        "parameters": [
          {
            "name": "minimo",
            "in": "query",
            "description": "Stock maximo para considerar un producto como bajo",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          },
          {
            "name": "page",
            "in": "query",
            "description": "Numero de pagina a consultar, comenzando desde 0",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Cantidad de registros por pagina",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          }
        ],
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "200": {
            "description": "Productos con stock bajo consultados correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelProductoDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/productos/disponibles": {
      "get": {
        "tags": [
          "Producto"
        ],
        "summary": "Consulta productos con stock disponible para clientes",
        "description": "Retorna productos cuyo stock es mayor a cero, incluyendo enlaces HATEOAS para navegar al recurso y a la coleccion.",
        "operationId": "listarProductosDisponibles",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Numero de pagina a consultar, comenzando desde 0",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Cantidad de registros por pagina",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          }
        ],
        "responses": {
          "500": {
            "description": "Error interno del servidor"
          },
          "200": {
            "description": "Productos disponibles consultados correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelProductoDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "precio": {
            "type": "number",
            "format": "double"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          }
        }
      },
      "EntityModelProductoDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Identificador unico del producto",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre comercial del producto",
            "example": "Arroz grano largo 1kg"
          },
          "precio": {
            "type": "number",
            "format": "double",
            "description": "Precio de venta del producto en pesos chilenos",
            "example": 1490
          },
          "stock": {
            "type": "integer",
            "format": "int32",
            "description": "Cantidad disponible en stock",
            "example": 80
          },
          "categoriaId": {
            "type": "integer",
            "format": "int64",
            "description": "Identificador de la categoria asociada",
            "example": 1
          },
          "categoriaNombre": {
            "type": "string",
            "description": "Nombre de la categoria asociada",
            "example": "Abarrotes"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "CollectionModelEntityModelProductoDTO": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "productos": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelProductoDTO"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}